mypy fixes

+++ tests/test_deployments/test_app.py
@@ -476,7 +476,8 @@ class TestTendermintBufferFailing(BaseTendermintServerTest):
     def teardown_class(cls) -> None:
         """Teardown the test."""
         # After the test, the node has hanged. We need to kill it and not stop it.
-        cls.tendermint_node._process.kill()
+        if cls.tendermint_node._process:
+            cls.tendermint_node._process.kill()
         cls.app_context.pop()
         shutil.rmtree(cls.tm_home, ignore_errors=True, onerror=readonly_handler)
